reorganize folders for different types of clients
