+++ client/benches/benchmark.rs
@@ -18,7 +18,7 @@ use cfxcore::{
     vm::{Env, Spec},
     vm_factory::VmFactory,
 };
-use client::{Client, Configuration};
+use client::{ArchiveClient, Configuration};
 use criterion::Criterion;
 use keylib::{Generator, KeyPair, Random};
 use parking_lot::{Condvar, Mutex};
@@ -29,7 +29,7 @@ fn txgen_benchmark(c: &mut Criterion) {
     let mut conf = Configuration::default();
     conf.raw_conf.test_mode = true;
     let exit = Arc::new((Mutex::new(false), Condvar::new()));
-    let handler = Client::start(conf, exit.clone()).unwrap();
+    let handler = ArchiveClient::start(conf, exit.clone()).unwrap();
     c.bench_function("Randomly generate 1 transaction", move |b| {
         b.iter(|| {
             handler.txgen.generate_transaction();
@@ -41,7 +41,7 @@ fn txexe_benchmark(c: &mut Criterion) {
     let mut conf = Configuration::default();
     conf.raw_conf.test_mode = true;
     let exit = Arc::new((Mutex::new(false), Condvar::new()));
-    let handler = Client::start(conf, exit.clone()).unwrap();
+    let handler = ArchiveClient::start(conf, exit.clone()).unwrap();
     let kp = KeyPair::from_secret(
         "46b9e861b63d3509c88b7817275a30d22d62c8cd8fa6486ddee35ef0d8e0495f"
             .parse()
